Refactor: fixtures/kitchen-sink.ts -> kitchen-sink.ts
